requirements.txt
asgiref==3.7.2
boto3==1.26.146
botocore==1.29.146
dj-database-url==2.0.0
Django==4.2.1
django-storages==1.13.2
djangorestframework==3.14.0
jmespath==1.0.1
psycopg==3.1.9
psycopg2-binary==2.9.6
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.3
urllib3==1.26.16
whitenoise==6.4.0
